refactor(api): rename and remove api files
